add generateStaticParams to blockexplorer address and txHash page (#825)
